+++ bomplatform-web/src/main/java/com/javabom/bomplatform/web/progressmission/service/ProgressMissionService.java
@@ -4,10 +4,10 @@ import com.javabom.bomplatform.core.mission.model.Mission;
 import com.javabom.bomplatform.core.progressmission.model.MissionReviewer;
 import com.javabom.bomplatform.core.user.model.User;
 import com.javabom.bomplatform.github.business.GithubBusiness;
-import com.javabom.bomplatform.web.progressmission.business.MissionBusiness;
-import com.javabom.bomplatform.web.progressmission.business.PickMissionReviewerBusiness;
-import com.javabom.bomplatform.web.progressmission.business.ProgressMissionBusiness;
-import com.javabom.bomplatform.web.progressmission.business.UserBusiness;
+import com.javabom.bomplatform.core.progressmission.business.MissionBusiness;
+import com.javabom.bomplatform.core.progressmission.business.PickMissionReviewerBusiness;
+import com.javabom.bomplatform.core.progressmission.business.ProgressMissionBusiness;
+import com.javabom.bomplatform.core.progressmission.business.UserBusiness;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
