+++ bomplatform-web/src/main/java/com/javabom/bomplatform/web/review/service/ReviewService.java
@@ -2,8 +2,8 @@ package com.javabom.bomplatform.web.review.service;
 
 import com.javabom.bomplatform.core.progressmission.model.ProgressMission;
 import com.javabom.bomplatform.github.business.GithubBusiness;
-import com.javabom.bomplatform.web.progressmission.business.ProgressMissionBusiness;
-import com.javabom.bomplatform.web.review.business.ReviewBusiness;
+import com.javabom.bomplatform.core.progressmission.business.ProgressMissionBusiness;
+import com.javabom.bomplatform.core.review.business.ReviewBusiness;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
