+++ bomplatform-web/src/test/java/com/javabom/bomplatform/web/progressmission/business/UserBusinessTest.java
@@ -1,5 +1,6 @@
 package com.javabom.bomplatform.web.progressmission.business;
 
+import com.javabom.bomplatform.core.progressmission.business.UserBusiness;
 import com.javabom.bomplatform.core.user.model.User;
 import com.javabom.bomplatform.core.user.model.UserRole;
 import com.javabom.bomplatform.core.user.repository.UserRepository;
